fix: refactor event source default naming
