+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/event/source/EventSource.java
@@ -20,8 +20,5 @@ public interface EventSource extends LifecycleAware {
    */
   void setEventHandler(EventHandler handler);
 
-  static String defaultNameFor(EventSource source) {
-    // we can have multiple event sources for the same class
-    return source.getClass().getName() + "#" + source.hashCode();
-  }
+
 }
