+++ operator-framework-core/src/test/java/io/javaoperatorsdk/operator/api/reconciler/EventSourceInitializerTest.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.processing.event.source;
+package io.javaoperatorsdk.operator.api.reconciler;
 
 import java.util.HashMap;
 
@@ -7,17 +7,17 @@ import org.junit.jupiter.api.Test;
 import io.javaoperatorsdk.operator.processing.event.source.polling.PollingEventSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
-class EventSourceTest {
+class EventSourceInitializerTest {
 
   @Test
   void defaultNameDifferentForOtherInstance() {
     var eventSource1 = new PollingEventSource(() -> new HashMap<>(), 1000, String.class);
     var eventSource2 = new PollingEventSource(() -> new HashMap<>(), 1000, String.class);
-    var eventSourceName1 = EventSource.defaultNameFor(eventSource1);
-    var eventSourceName2 = EventSource.defaultNameFor(eventSource2);
+    var eventSourceName1 = EventSourceInitializer.generateNameFor(eventSource1);
+    var eventSourceName2 = EventSourceInitializer.generateNameFor(eventSource2);
 
     assertThat(eventSourceName1).isNotEqualTo(eventSourceName2);
   }
+
 }
