+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/sample/createupdateeventfilter/CreateUpdateEventFilterTestReconciler.java
@@ -106,7 +106,7 @@ public class CreateUpdateEventFilterTestReconciler
             .withLabelSelector("integrationtest = " + this.getClass().getSimpleName())
             .build();
     informerEventSource = new InformerEventSource<>(informerConfiguration, client);
-    return EventSourceInitializer.defaultNamedEventSources(informerEventSource);
+    return EventSourceInitializer.generateNameForEventSources(informerEventSource);
   }
 
   @Override
