+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/sample/informereventsource/InformerEventSourceTestCustomReconciler.java
@@ -47,7 +47,7 @@ public class InformerEventSourceTestCustomReconciler
             .build();
 
     return EventSourceInitializer
-        .defaultNamedEventSources(new InformerEventSource<>(config, context));
+        .generateNameForEventSources(new InformerEventSource<>(config, context));
   }
 
   @Override
