+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/sample/standalonedependent/StandaloneDependentTestReconciler.java
@@ -40,7 +40,7 @@ public class StandaloneDependentTestReconciler
   public Map<String, EventSource> prepareEventSources(
       EventSourceContext<StandaloneDependentTestCustomResource> context) {
     return EventSourceInitializer
-        .defaultNamedEventSources(deploymentDependent.initEventSource(context));
+        .generateNameForEventSources(deploymentDependent.initEventSource(context));
   }
 
   @Override
