+++ sample-operators/tomcat-operator/src/main/java/io/javaoperatorsdk/operator/sample/WebappReconciler.java
@@ -73,7 +73,7 @@ public class WebappReconciler
             .withAssociatedSecondaryResourceIdentifier(tomcatFromWebAppSpec)
             .build();
     return EventSourceInitializer
-        .defaultNamedEventSources(new InformerEventSource<>(configuration, context));
+        .generateNameForEventSources(new InformerEventSource<>(configuration, context));
   }
 
   /**
