+++ sample-operators/webpage/src/main/java/io/javaoperatorsdk/operator/sample/WebPageReconciler.java
@@ -65,7 +65,7 @@ public class WebPageReconciler
         new InformerEventSource<>(InformerConfiguration.from(context, Service.class)
             .withLabelSelector(LOW_LEVEL_LABEL_KEY)
             .build(), context);
-    return EventSourceInitializer.defaultNamedEventSources(configMapEventSource,
+    return EventSourceInitializer.generateNameForEventSources(configMapEventSource,
         deploymentEventSource,
         serviceEventSource);
   }
